// common/types/Person.ts
export interface PersonTypes {
  id?: number;
  firstName: string;
  lastName: string;
  placeOfBirth?: string;
  dateOfBirth?: string;
  gender?: string;
  religion?: string;
  bloodType?: string;
  maritalStatus?: string;
  address?: string;
  rt?: string;
  rw?: string;
  village?: string;
  district?: string;
  city?: string;
  province?: string;
  nationality?: string;
  postalCode?: string;
  phoneNumber?: string;
  email?: string;
  idNumber?: string;
  occupation?: string;
  hobby?: string;
  createdAt?: Date;
  updatedAt?: Date;
  status?: boolean;
}